refactor: integration -> gil
