+++ gil/index.js
@@ -2,7 +2,7 @@ const CUDBatch = require('./sink/CUDBatch');
 const CUDCommand = require('./sink/CUDCommand');
 const CypherSink = require('./sink/CypherSink');
 const DataSink = require('./sink/DataSink');
-const neo4j = require('./driver/');
+const neo4j = require('./driver');
 
 module.exports = {
     CUDBatch, CUDCommand, CypherSink, DataSink, neo4j,
