+++ gil/sink/CUDBatch.js
