+++ gil/sink/CUDCommand.js
