+++ gil/sink/CypherSink.test.js
