+++ gil/sink/Strategy.js
