+++ services/http/node.js
@@ -1,6 +1,6 @@
 const _ = require('lodash');
 const moment = require('moment');
-const neo4j = require('../../integration/driver');
+const neo4j = require('../../gil/driver');
 const Integer = require('neo4j-driver/lib/integer.js');
 const common = require('../common');
