+++ services/http/node.test.js
@@ -1,6 +1,6 @@
 const node = require('./node');
 const test = require('../../test');
-const integration = require('../../integration');
+const gil = require('../../gil');
 const sinon = require("sinon");
 
 describe('Node Function', () => {
@@ -25,12 +25,12 @@ describe('Node Function', () => {
             writeTransaction: f => f(tx) 
         });
 
-        driver = sinon.stub(integration.neo4j, 'getDriver');
+        driver = sinon.stub(gil.neo4j, 'getDriver');
         driver.returns({ session });
     });
 
     afterEach(() => {
-        integration.neo4j.getDriver.restore();
+        gil.neo4j.getDriver.restore();
     });
 
     it('should process many messages', () => {
