+++ services/pubsub/cypherPubsub.js
@@ -1,12 +1,12 @@
 const _ = require('lodash');
-const integration = require('../../integration');
+const gil = require('../../gil');
 
 // https://cloud.google.com/functions/docs/writing/background#function_parameters
 const cypher = (pubSubEvent, context, callback) => {
     const input = JSON.parse(Buffer.from(pubSubEvent.data, 'base64').toString());
 
-    const cypher = new integration.CypherSink(input);
-    const sink = new integration.DataSink([cypher]);
+    const cypher = new gil.CypherSink(input);
+    const sink = new gil.DataSink([cypher]);
     
     return sink.run()
         .then(results => callback(null, results))
